flight-alerts/requirements.txt
amadeus
python-dotenv
requests
psycopg2-binary
